Convert VM templates back to VMs during cleanup
In vSphere 6.0, you can encounter a situation where you are unable to
delete a VM template inside a folder and receive the following error:
"InvalidDeviceSpec: Invalid configuration for device '0'."

This is remedied by converting the template into an actual VM, which
then can be deleted.

The version of ruby used here was also extremely out of date and has
been bumped to reflect what we use elsewhere. This required explicitly
adding the xmlrpc gem, which got moved out of the Ruby standard library
in 2.4.

Signed-off-by: Kenneth Lakin <klakin@pivotal.io>